refactor: extract useSwapResultsContext
